examples/wikitext2/requirements.txt
torchtext
torchdata
fairscale
torchgpipe
transformers